Create BaseFixtureFactory to consolidate AutoDataAttribute logic
- Add BaseFixtureFactory class to centralize common fixture configuration
- Refactor LoggingAutoDataAttribute, MediatRAutoDataAttribute, and MediatRLambdaAutoDataAttribute to use factory pattern
- Remove duplicated fixture setup code (ThrowingRecursionBehavior removal, OmitOnRecursionBehavior addition, AutoNSubstituteCustomization)
- Each attribute now focuses only on its specific specimen builders and customizations
