+++ AlexaVoxCraft.TestKit/Attributes/LoggingAutoDataAttribute.cs
@@ -1,0 +1,19 @@
+using AlexaVoxCraft.TestKit.SpecimenBuilders;
+using AutoFixture;
+using AutoFixture.Xunit3;
+
+namespace AlexaVoxCraft.TestKit.Attributes;
+
+/// <summary>
+/// AutoData attribute that uses a customized fixture for consistent test data generation.
+/// </summary>
+public class LoggingAutoDataAttribute() : AutoDataAttribute(CreateFixture)
+{
+    private static IFixture CreateFixture()
+    {
+        return BaseFixtureFactory.CreateFixture(fixture =>
+        {
+            fixture.Customizations.Add(new LogEventSpecimenBuilder());
+        });
+    }
+}
