+++ AlexaVoxCraft.TestKit/RequestSpecifications/LambdaContextSpecification.cs
@@ -1,0 +1,7 @@
+using Amazon.Lambda.Core;
+
+namespace AlexaVoxCraft.TestKit.RequestSpecifications;
+
+public class LambdaContextSpecification : BaseTypedSpecification<ILambdaContext>
+{
+}
