+++ AlexaVoxCraft.TestKit/RequestSpecifications/LogEventSpecification.cs
@@ -1,0 +1,7 @@
+using Serilog.Events;
+
+namespace AlexaVoxCraft.TestKit.RequestSpecifications;
+
+public class LogEventSpecification : BaseTypedSpecification<LogEvent>
+{
+}
