+++ AlexaVoxCraft.TestKit/SpecimenBuilders/SkillRequestFactorySpecimenBuilder.cs
@@ -1,7 +1,5 @@
-using System.Reflection;
 using AlexaVoxCraft.MediatR;
 using AlexaVoxCraft.Model.Request;
-using AlexaVoxCraft.Model.Request.Type;
 using AlexaVoxCraft.TestKit.RequestSpecifications;
 using AutoFixture.Kernel;
 using NSubstitute;
