+++ test/AlexaVoxCraft.Logging.Tests/AlexaVoxCraft.Logging.Tests.csproj
@@ -25,16 +25,13 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="AutoFixture" Version="4.18.1" />
-    <PackageReference Include="AutoFixture.Xunit3" Version="4.19.0" />
-    <PackageReference Include="NSubstitute" Version="5.1.0" />
     <PackageReference Include="AwesomeAssertions" Version="9.0.0" />
     <PackageReference Include="Serilog" Version="4.3.0" />
     <PackageReference Include="Serilog.Sinks.TestCorrelator" Version="4.0.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\AlexaVoxCraft.Logging\AlexaVoxCraft.Logging.csproj" />
+    <ProjectReference Include="..\..\AlexaVoxCraft.TestKit\AlexaVoxCraft.TestKit.csproj" />
   </ItemGroup>
 
 </Project>
