+++ test/AlexaVoxCraft.Model.Tests/IntentTests.cs
@@ -1,5 +1,4 @@
 ﻿using AlexaVoxCraft.Model.Request;
-using Xunit;
 
 namespace AlexaVoxCraft.Model.Tests;
