+++ test/AlexaVoxCraft.Model.Tests/RequestTests.cs
@@ -3,7 +3,6 @@ using AlexaVoxCraft.Model.Request;
 using AlexaVoxCraft.Model.Request.Type;
 using AlexaVoxCraft.Model.Serialization;
 using AlexaVoxCraft.Model.Tests.Extensions;
-using Xunit;
 
 namespace AlexaVoxCraft.Model.Tests;
